refactor: pro.ts
