+++ src/pro.ts
@@ -1,0 +1,7 @@
+export const suffix = (distribution: string): string => {
+  return isPro(distribution) ? '-pro' : '';
+};
+
+export const isPro = (distribution: string): boolean => {
+  return distribution === 'goreleaser-pro';
+};
